cast and data loss warnings clear

+++ src/scenario/scenario.c
@@ -210,7 +210,7 @@ int scenario_get_state_buffer_size_by_savegame_version(int savegame_version)
     } else {
         calculate_buffer_offsets(SCENARIO_CURRENT_VERSION);
     }
-    return buffer_offsets.end;
+    return (int) buffer_offsets.end;
 }
 
 int scenario_get_state_buffer_size_by_scenario_version(int scenario_version)
@@ -221,7 +221,7 @@ int scenario_get_state_buffer_size_by_scenario_version(int scenario_version)
         return 1830;
     } else {
         calculate_buffer_offsets(scenario_version);
-        return buffer_offsets.end;
+        return (int) buffer_offsets.end;
     }
 }
 
@@ -549,13 +549,13 @@ void scenario_load_state(buffer *buf, int version)
     }
 
     if (version > SCENARIO_LAST_NO_CUSTOM_VARIABLES && version <= SCENARIO_LAST_STATIC_ORIGINAL_DATA) {
-        buffer_set(buf, buffer_offsets.custom_variables);
+        buffer_set(buf, (int) buffer_offsets.custom_variables);
         scenario_custom_variable_load_state_old_version(buf);
     } else {
         scenario_custom_variable_delete_all();
     }
 
-    buffer_set(buf, buffer_offsets.custom_name);
+    buffer_set(buf, (int) buffer_offsets.custom_name);
     if (version > SCENARIO_LAST_UNVERSIONED) {
         buffer_read_raw(buf, scenario.empire.custom_name, sizeof(scenario.empire.custom_name));
     }
@@ -573,7 +573,7 @@ void scenario_load_state(buffer *buf, int version)
 void scenario_description_from_buffer(buffer *buf, uint8_t *description, int version)
 {
     calculate_buffer_offsets(version);
-    buffer_set(buf, buffer_offsets.briefing);
+    buffer_set(buf, (int) buffer_offsets.briefing);
     buffer_read_raw(buf, description, MAX_BRIEF_DESCRIPTION);
 }
 
@@ -585,7 +585,7 @@ int scenario_climate_from_buffer(buffer *buf, int version)
     } else if (version <= SCENARIO_LAST_NO_WAGE_LIMITS) {
         buffer_set(buf, 1764);
     } else {
-        buffer_set(buf, buffer_offsets.misc + 36); // measure offsets to make sure they match
+        buffer_set(buf, (int) buffer_offsets.misc + 36); // measure offsets to make sure they match
     }
     return buffer_read_u8(buf);
 }
@@ -598,7 +598,7 @@ int scenario_invasions_from_buffer(buffer *buf, int version)
         return scenario_invasion_count_active_from_buffer(buf);
     } else {
         calculate_buffer_offsets(version);
-        buffer_set(buf, buffer_offsets.original_invasions_part1 + (MAX_ORIGINAL_INVASIONS * 2));
+        buffer_set(buf, (int) buffer_offsets.original_invasions_part1 + (MAX_ORIGINAL_INVASIONS * 2));
         for (int i = 0; i < MAX_ORIGINAL_INVASIONS; i++) {
             if (buffer_read_i16(buf)) {
                 num_invasions++;
@@ -611,21 +611,21 @@ int scenario_invasions_from_buffer(buffer *buf, int version)
 int scenario_image_id_from_buffer(buffer *buf, int version)
 {
     calculate_buffer_offsets(version);
-    buffer_set(buf, buffer_offsets.image);
+    buffer_set(buf, (int) buffer_offsets.image);
     return buffer_read_i16(buf);
 }
 
 int scenario_rank_from_buffer(buffer *buf, int version)
 {
     calculate_buffer_offsets(version);
-    buffer_set(buf, buffer_offsets.image + 4);
+    buffer_set(buf, (int) buffer_offsets.image + 4);
     return buffer_read_i16(buf);
 }
 
 void scenario_open_play_info_from_buffer(buffer *buf, int version, int *is_open_play, int *open_play_id)
 {
     calculate_buffer_offsets(version);
-    buffer_set(buf, buffer_offsets.image + 2);
+    buffer_set(buf, (int) buffer_offsets.image + 2);
     *is_open_play = buffer_read_i16(buf);
 
     if (version <= SCENARIO_LAST_UNVERSIONED) {
@@ -633,7 +633,7 @@ void scenario_open_play_info_from_buffer(buffer *buf, int version, int *is_open_
     } else if (version <= SCENARIO_LAST_NO_WAGE_LIMITS) {
         buffer_set(buf, 1778);
     } else {
-        buffer_set(buf, buffer_offsets.misc + 50);
+        buffer_set(buf, (int) buffer_offsets.misc + 50);
     }
     *open_play_id = buffer_read_u8(buf);
 }
@@ -641,7 +641,7 @@ void scenario_open_play_info_from_buffer(buffer *buf, int version, int *is_open_
 int scenario_start_year_from_buffer(buffer *buf, int version)
 {
     calculate_buffer_offsets(version);
-    buffer_set(buf, buffer_offsets.start_info);
+    buffer_set(buf, (int) buffer_offsets.start_info);
     return buffer_read_i16(buf);
 }
 
@@ -653,7 +653,7 @@ void scenario_objectives_from_buffer(buffer *buf, int version, scenario_win_crit
     } else if (version <= SCENARIO_LAST_NO_WAGE_LIMITS) {
         buffer_set(buf, 1632);
     } else {
-        buffer_set(buf, buffer_offsets.win_criteria + 2); // 2 bits for favour reset
+        buffer_set(buf, (int) buffer_offsets.win_criteria + 2); // 2 bits for favour reset
     }
     win_criteria->culture.goal = buffer_read_i32(buf);
     win_criteria->prosperity.goal = buffer_read_i32(buf);
@@ -675,7 +675,7 @@ void scenario_objectives_from_buffer(buffer *buf, int version, scenario_win_crit
 void scenario_map_data_from_buffer(buffer *buf, int *width, int *height, int *grid_start, int *grid_border_size, int version)
 {
     calculate_buffer_offsets(version);
-    buffer_set(buf, buffer_offsets.map_size);
+    buffer_set(buf, (int) buffer_offsets.map_size);
     *width = buffer_read_i32(buf);
     *height = buffer_read_i32(buf);
     *grid_start = buffer_read_i32(buf);
